+++ R/yaml_to_aeme.R
@@ -74,7 +74,8 @@ yaml_to_aeme <- function(path, file) {
   if (!is.null(yaml$observations$lake)) {
     yaml$observations$lake <- read.csv(file.path(path,
                                                  yaml$observations$lake)) |>
-      dplyr::mutate(Date = as.Date(datetime))
+      dplyr::mutate(Date = as.Date(datetime)) |>
+      normalise_lake_obs()
   }
   if (!is.null(yaml$observations$level)) {
     yaml$observations$level <- read.csv(file.path(path,